Code refactor KT-1
