+++ package.json
@@ -25,6 +25,8 @@
         "eslint": "^7.2.0",
         "react": "^17.0.1",
         "react-dom": "^17.0.1",
+        "react-redux": "^7.1.1",
+        "redux": "^4.0.4",
         "typescript": "^4.2.4"
     },
     "prettier": {
