+++ src/core/models/exchange-rates.ts
@@ -1,4 +1,6 @@
-/** Currency and their exchange rate. */
-export interface ExchangeRates {
-    readonly [key: string]: number;
-}
+import {CurrencyIso} from 'core/utils/currencyList';
+
+/** Валюты и их обменный курс. */
+export type ExchangeRates = {
+    readonly [key in CurrencyIso]: number;
+};
